Updated DAWSON search logic to use lower DAWSON environments if website is running in development, QA, or training environments. Consolidated logic for determining DAWSON environments into a single location.
