+++ website/search/views.py
@@ -13,7 +13,7 @@ from home.models.snippets.judges import JudgeProfile
 from django.db.models import Q
 
 from search.dawson import is_docket_number
-from search.dawson_client import get_case_record, dawson_public_site_url
+from search.dawson_client import get_case_record, get_environment_specific_dawson_url
 from search.models.definitionsQuery import DefinitionsQuery
 
 # from search.models import DefinitionsQuery
@@ -240,7 +240,7 @@ def search(request):
                 "value": {
                     "heading": "Looking for a docket number?",
                     "text": RichText(
-                        f'<p data-block-key="5ybz3">Docket numbers must be entered in the format <b>123-19</b>.</p><p data-block-key="7negh"><b>Example</b>: “Docket Number 123-19”, “Docket No. 123-19”, or “123-19”</p><p data-block-key="fjj0k" class="right-aligned"><a class="dawson-search-link" href="{dawson_public_site_url()}" target="_blank" rel="noopener">Search DAWSON’s Docket Records<span class="launch-icon" aria-hidden="true"></span></a></p>'
+                        f'<p data-block-key="5ybz3">Docket numbers must be entered in the format <b>123-19</b>.</p><p data-block-key="7negh"><b>Example</b>: “Docket Number 123-19”, “Docket No. 123-19”, or “123-19”</p><p data-block-key="fjj0k" class="right-aligned"><a class="dawson-search-link" href="{get_environment_specific_dawson_url()}" target="_blank" rel="noopener">Search DAWSON’s Docket Records<span class="launch-icon" aria-hidden="true"></span></a></p>'
                     ),
                     "callout_type": "warning",
                 },
